pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "distributedlog"
version = "0.1.0"
description = "A distributed commit log system built from scratch"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Horace Njoroge"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "pyyaml>=6.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
    "aiofiles>=23.2.1",
    "crc32c>=2.3",           # Fast CRC for checksums
    "msgpack>=1.0.7",        # Efficient binary serialization
    "click>=8.1.7",          # CLI framework
    "rich>=13.7.0",          # Beautiful terminal output
    "tenacity>=8.2.3",       # Retry logic
    "watchdog>=3.0.0",       # File system monitoring
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "hypothesis>=6.92.0",    # Property-based testing
    "mypy>=1.8.0",
    "ruff>=0.1.9",
    "black>=23.12.0",
    "pre-commit>=3.6.0",
    "grpcio-testing>=1.60.0",
]

[project.scripts]
dlog = "distributedlog.admin.cli:main"
dlog-broker = "distributedlog.broker.server.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["distributedlog*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["distributedlog/tests"]
addopts = "-v --tb=short"
timeout = 30

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "A", "C4", "PT", "RET", "SIM"]
ignore = ["ANN101", "ANN102"]

[tool.black]
line-length = 100
target-version = ["py310"]
